restructure into comp module
