+++ ogma/src/eng/ty/mod.rs
@@ -2,4 +2,3 @@
 use super::*;
 
 mod annotate;
-mod infer;
